+++ .gitignore
@@ -10,8 +10,9 @@ video0.mp4
 what.html
 ffmpeg
 ytdl-core.zip
-birdycopy.js
 fetch.js
 ytdl-coresomething.zip
 gru
 discord.js(pull request)
+ytsr
+ytsr
