+++ commands/assets/sprite.png
