+++ commands/status.js
@@ -5,7 +5,7 @@ module.exports = {
 	name: 'status',
 	description: 'status function that you can reload to update Birdy7\'s status',
 	fetchStatus() {
-		const activityupdate = null;
+		const activityupdate = 'Internet is a being slow';
 		const activityname = '>help';
 		const BirdyActivity = activityupdate == undefined ? `${activityname}` : activityupdate == null ? `${activityname}` : `${activityname} | ${activityupdate}`;
 		return BirdyActivity;
